# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Tailor CRM API
  version: 3.0.0
  description: |
    # Production-Ready REST API for Tailor Management SaaS

    **Last Updated:** May 10, 2026
    **Status:** Fully Implemented & Tested
    **Total Endpoints:** 80 (73 + 7 system)

    ## Response Format (ALL endpoints)

    Success:   { success: true,  message: "...", data: {} }
    Error:     { success: false, message: "...", error: { code: "...", details: {} } }
    Paginated: { success: true,  message: "...", data: { items: [], pagination: { page, limit, total } } }


    ## Field Naming Convention
    - Requests:  snake_case  (customer_id, delivery_date)
    - Responses: camelCase   (customerId, deliveryDate)

    ## Order Status Flow
    pending → in_progress → ready → delivered  (cannot skip)

    ## Error Codes
    INVALID_INPUT | UNAUTHORIZED | FORBIDDEN | NOT_FOUND
    DUPLICATE_PHONE | INVALID_STATUS_TRANSITION
    CUSTOMER_NOT_FOUND | ORDER_NOT_FOUND | SHOP_NOT_FOUND
    MEASUREMENT_NOT_FOUND | STAFF_NOT_FOUND | INVOICE_NOT_FOUND
    INTERNAL_ERROR | RATE_LIMIT_EXCEEDED | SERVICE_UNAVAILABLE

    ## Rate Limiting
    All requests are rate-limited per IP address:
    - **Global**: 100 requests / 15 minutes (default)
    - **Login**: 5 requests / 15 minutes
    - **OTP/Verify**: 3 requests / 15 minutes
    - **Refresh**: 10 requests / 15 minutes
    - **Upload**: 20 requests / 15 minutes
    
    Rate limit status in response headers:
    - `RateLimit-Limit`: Maximum requests allowed
...

...

...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        ## JWT Authentication with Token Rotation & Blacklist
        
        **Obtain tokens from:**
        - POST /auth/login (Firebase production)
        - POST /auth/login-test (Development only)
        
        **Token Structure:**
        - Access Token: 7-day expiry, contains userId, phone, shopId, sessionId, jti
        - Refresh Token: Long-lived, used for rotation
        
        **Token Rotation Behavior:**
        - Every refresh returns NEW access + refresh tokens
        - Old refresh token becomes invalid (single-use)
        - Prevents token reuse attacks
        - Session ID maintained across rotations
        
        **Security Features:**
        - Redis-backed session tracking
        - Access token blacklist on logout
        - Session activity validation
        - JTI (JWT ID) claim prevents replay attacks
        - Automatic revocation on suspicious activity
        
        **On 401 Unauthorized:**
        - Token expired? Call POST /auth/refresh-token
        - Still fails? Re-login with POST /auth/login
        - Token revoked? Session blacklisted, must re-login
        
        **Usage:**
        Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
...
security:
  - BearerAuth: []
...
paths:
...
  /measurement:
    post:
      tags: [Measurement]
      summary: Create measurement record
      description: All measurement values stored as JSONB. Any field names allowed. Decimals up to 2 places.
      operationId: createMeasurement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [customer_id, measurements_data]
              properties:
                customer_id:
                  type: integer
                outfit_type:
                  type: string
                outfit_label:
                  type: string
                measurements_data:
                  type: object
                  additionalProperties:
                    type: number
            example:
              customer_id: 1
              outfit_type: "pants"
              outfit_label: "Pants"
              measurements_data:
                Length: 42.5
                "Waist Bottom": 32.0
                "Hip Circumference Bottom": 38.0
                "Thigh Circumference": 22.0
                "Knee Circumference": 16.0
                Crotch: 12.0
                Fly: 8.0
                "Ankle Circumference": 14.0
      responses:
        '201':
          description: Measurement created
          content:
            application/json:
              example:
                success: true
                message: Measurement created successfully
                data:
                  id: 1
                  customerId: 1
                  outfitType: "pants"
                  outfitLabel: "Pants"
                  measurementsData:
                    Length: 42.5
                    "Waist Bottom": 32.0
                    "Hip Circumference Bottom": 38.0
                    "Thigh Circumference": 22.0
                    "Knee Circumference": 16.0
                    Crotch: 12.0
                    Fly: 8.0
                    "Ankle Circumference": 14.0
                  createdAt: "2026-04-29T10:00:00Z"
                  updatedAt: "2026-04-29T10:00:00Z"
  /measurement/customer/{customerId}:
    get:
      tags: [Measurement]
      summary: Get all measurements for a customer
      operationId: getMeasurementsByCustomer
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: integer
          example: 1
        - name: outfit_type
          in: query
          required: false
          description: Filter measurements by outfit type (shirt, pant, kurta, etc.)
          schema:
            type: string
          example: shirt
      responses:
        '200':
          description: Measurements retrieved
          content:
            application/json:
              example:
                success: true
                message: Measurements retrieved successfully
                data:
                  items:
                    - id: 1
                      customerId: 1
                      outfitType: "pants"
                      outfitLabel: "Pants"
                      measurementsData:
                        Length: 42.5
                        "Waist Bottom": 32.0
                      createdAt: "2026-04-29T10:00:00Z"
                    - id: 2
                      customerId: 1
                      outfitType: "shirt"
                      outfitLabel: "Shirt"
                      measurementsData:
                        "Medium Length": 28.0
                        Chest: 40.0
                        Waist: 34.0
                      createdAt: "2026-04-20T10:00:00Z"
  /measurement/{id}:
    get:
      tags: [Measurement]
      summary: Get measurement by ID
      operationId: getMeasurement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Measurement retrieved
          content:
            application/json:
              example:
                success: true
                message: Measurement retrieved successfully
                data:
                  id: 1
                  customerId: 1
                  outfitType: "pants"
                  outfitLabel: "Pants"
                  measurementsData:
                    Length: 42.5
                    "Waist Bottom": 32.0
                  createdAt: "2026-04-29T10:00:00Z"
        '404':
          description: Measurement not found
          content:
            application/json:
              example:
                success: false
                message: Measurement not found
                error:
                  code: MEASUREMENT_NOT_FOUND

    put:
      tags: [Measurement]
      summary: Update measurement
      operationId: updateMeasurement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                measurements_data:
                  type: object
                  additionalProperties:
                    type: number
            example:
              measurements_data:
                Length: 43.0
                "Waist Bottom": 33.0
                "Hip Circumference Bottom": 39.0
      responses:
        '200':
          description: Measurement updated
          content:
            application/json:
              example:
                success: true
                message: Measurement updated successfully
                data:
                  id: 1
                  measurementsData:
                    Length: 43.0
                    "Waist Bottom": 33.0
                    "Hip Circumference Bottom": 39.0
                  updatedAt: "2026-04-29T14:00:00Z"

    delete:
      tags: [Measurement]
      summary: Delete measurement
      operationId: deleteMeasurement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Measurement deleted
          content:
            application/json:
              example:
                success: true
                message: Measurement deleted successfully
                data: null

  # ══════════════════════════════════════════════════════
  # PAYMENT — 3 endpoints
  # ══════════════════════════════════════════════════════
...